reorganizie file
change adpater location, create a whole directory for activities
