+++ app/src/androidTest/java/com/github/multimatum_team/multimatum/QRCodeReaderActivityTest.kt
@@ -7,6 +7,7 @@ import androidx.test.espresso.assertion.ViewAssertions
 import androidx.test.espresso.intent.Intents
 import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.rule.GrantPermissionRule
+import com.github.multimatum_team.multimatum.activity.QRCodeReaderActivity
 import org.junit.After
 import org.junit.Before
 import org.junit.Rule
