+++ app/src/main/java/com/github/multimatum_team/multimatum/activity/QRCodeReaderActivity.kt
@@ -1,4 +1,4 @@
-package com.github.multimatum_team.multimatum
+package com.github.multimatum_team.multimatum.activity
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
@@ -9,6 +9,7 @@ import com.budiyev.android.codescanner.CodeScannerView
 import com.budiyev.android.codescanner.DecodeCallback
 import com.budiyev.android.codescanner.ErrorCallback
 import com.budiyev.android.codescanner.ScanMode
+import com.github.multimatum_team.multimatum.R
 
 /**
  * The purpose of this activity is to provide the user an interface to scan QR-Codes.
