+++ app/src/main/res/layout/activity_deadline_details.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DeadlineDetailsActivity">
+    tools:context=".activity.DeadlineDetailsActivity">
 
     <EditText
         android:id="@+id/deadline_details_activity_title"
