+++ app/src/main/res/layout/activity_profile.xml
@@ -5,7 +5,7 @@
     android:id="@+id/profile_frag"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ProfileActivity">
+    tools:context=".activity.ProfileActivity">
 
     <TextView
         android:id="@+id/login_message"
