+++ app/src/main/res/layout/activity_qrgenerator.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".QRGeneratorActivity">
+    tools:context=".activity.QRGeneratorActivity">
 
     <ImageView
         android:id="@+id/QRGenerated"
