+++ app/src/main/res/layout/activity_signin.xml
@@ -5,7 +5,7 @@
     android:id="@+id/account_activity"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SignInActivity">
+    tools:context=".activity.SignInActivity">
 
     <Button
         android:id="@+id/sign_in_button"
